supportsConnect/requirements.txt
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
WTForms==2.3.3
